Add accounts, offline-first sync and the Universal Dhikr board
Everything here is additive. No existing screen, table, preference or feature
changed behaviour, and the 50 tests from v1.0.0 still pass untouched.

Sync. Every count now writes the day ledger and an append-only outbox entry in
one transaction — a count that reached the device but not the queue would never
reach the cloud, so neither commits without the other. Each operation carries a
client-generated UUID that becomes its remote document id, and the aggregation
function claims a marker inside the same transaction that moves the totals, so a
client retry, a function retry and a rewritten document all converge and stop.
Nothing is ever discarded: a failed push increments an attempt counter and stays
queued, and no code path deletes an unsynced operation.

Existing history is claimed exactly once, as a single operation whose id derives
from the device, sized as the local lifetime total minus everything the outbox
has ever carried. That is precisely the counting that predates this feature, so
12,500 dhikr before signing in are 12,500 after it.

Accounts. Google via Credential Manager and Facebook via its Login SDK, both
ending at one Firebase uid. Signing out clears the session and nothing else.

Universal Dhikr. A new tab showing the worldwide total, this account's
contribution and its share — derived from both figures on every frame, never
stored. The contribution is the server's total plus what is still queued
locally, so it neither stalls mid-sync nor understates a second device.

Firebase is initialised programmatically rather than through the google-services
plugin, which fails the build outright without its config file. With no
credentials the app compiles and behaves exactly as v1.0.0 did, and the board
says so instead of erroring.

Backend: Firestore rules closing totals to every client, aggregation functions,
a nightly reconciliation pass, and setup instructions.

31 new tests. Verified the release APK keeps the v1.0.0 package id and signature
and raises versionCode, so it installs as an update.

+++ app/src/main/java/com/lillah/dhikr/data/local/entity/Entities.kt
@@ -120,7 +120,8 @@ data class SyncOperationEntity(
     val dhikrId: Long? = null,
     val dhikrName: String? = null,
     val epochDay: Long,
-    val delta: Int,
+    /** Long, not Int: a baseline claim carries a whole history in one operation. */
+    val delta: Long,
     val createdAt: Long,
     /** Matches a [com.lillah.dhikr.domain.sync.SyncState] name. */
     val state: String,
